Verification (#52)
* Move getMuteRoleId into a utils module

* Move logic for applying a mute role to utils

* Handle missing DM forward channel

* Update drizzle journals

* Don't delete original reply if it's already ephemeral

* Add verification commands

* Notify the moderator when a verifications' time runs out

+++ apps/bot/src/util/errorFollowUp.ts
@@ -3,13 +3,16 @@ import type { ChatInputCommandInteraction } from "discord.js";
 /**
  * Send an error follow-up message to the user
  *
- * Deletes the original reply and sends a new ephemeral message
+ * Deletes the original reply if it isn't ephemeral
+ * and sends a new ephemeral message
  */
 export const errorFollowUp = async (
   itx: ChatInputCommandInteraction,
   content: string,
 ) => {
-  await itx.deleteReply();
+  if (itx.deferred && !itx.ephemeral) {
+    await itx.deleteReply();
+  }
   await itx.followUp({
     content,
     ephemeral: true,
